requirements.txt
# ==========================================
# PDF Consultor - Requirements (FAISS)
# ==========================================
# Sistema RAPTOR (Reciprocal Rank Fusion + Abstração Hierárquica)
# ==========================================

# Web Framework & API
fastapi==0.129.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.22
python-dotenv>=1.0.0

# PDF Processing
pypdf>=6.7.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.13

# Vector Store & Search (FAISS em vez de ChromaDB)
faiss-cpu>=1.7.4
numpy>=1.26.0

# Embeddings & ML
sentence-transformers>=2.7.0
torch>=1.11.0
scikit-learn>=1.1.0

# NLP & Text Processing
nltk>=3.9.0
tiktoken>=0.12.0

# LLM & RAG
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.3.0

# Utilities
loguru>=0.7.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
